moved function from frustum to camera
